refactor to apply DRY in view components
