+++ src/renderer/components/ModeToggle.tsx
@@ -8,7 +8,7 @@ import {
   DropdownMenuItem,
 } from '../shad/ui/dropdown-menu';
 
-export function ModeToggle() {
+export const ModeToggle: React.FC = () => {
   const { setTheme } = useTheme();
 
   return (
@@ -33,4 +33,4 @@ export function ModeToggle() {
       </DropdownMenuContent>
     </DropdownMenu>
   );
-}
+};
