+++ src/renderer/routes.tsx
@@ -55,8 +55,8 @@ export default function appRoutes() {
 function RequireAuth({ children }: PropsWithChildren) {
   const { authed } = useAuth();
 
-  // if (authed) {
-  if (authed || process.env.NODE_ENV === 'development') {
+  if (authed) {
+    // if (authed || process.env.NODE_ENV === 'development') {
     return (
       <>
         {children}
