+++ src/renderer/views/Journals/index.tsx
@@ -14,16 +14,12 @@ import {
 import { Separator } from 'renderer/shad/ui/separator';
 import { Table, TableBody, TableCell, TableRow } from 'renderer/shad/ui/table';
 import { defaultSortingFunctions } from 'renderer/lib/utils';
-import type { Journal } from 'types';
+import type { HasMiniView, Journal } from 'types';
 import { toString } from 'lodash';
 
 export type JournalView = Journal & { amount: number };
 
-interface JournalPageProps {
-  isMini?: boolean;
-}
-
-const JournalsPage: React.FC<JournalPageProps> = ({ isMini = false }) => {
+const JournalsPage: React.FC<HasMiniView> = ({ isMini = false }) => {
   console.log('JournalsPage', isMini);
   const [journals, setJounrals] = useState<JournalView[]>([]);
   const [filteredJournals, setFilteredJournals] = useState<JournalView[]>(
@@ -80,22 +76,21 @@ const JournalsPage: React.FC<JournalPageProps> = ({ isMini = false }) => {
     [journals],
   );
 
-  useEffect(
-    () =>
-      void (async () =>
-        setJounrals(
-          ((await window.electron.getJournals()) as Journal[]).map(
-            (journal) => ({
-              ...journal,
-              amount: journal.journalEntries.reduce(
-                (acc, entry) => acc + entry.debitAmount,
-                0,
-              ),
-            }),
-          ),
-        ))(),
-    [],
-  );
+  useEffect(() => {
+    const fetchJournals = async () => {
+      const journals = (await window.electron.getJournals()) as Journal[];
+      const updatedJournals = journals.map((journal) => ({
+        ...journal,
+        amount: journal.journalEntries.reduce(
+          (acc, entry) => acc + entry.debitAmount,
+          0,
+        ),
+      }));
+      setJounrals(updatedJournals);
+    };
+
+    fetchJournals();
+  }, []);
 
   useEffect(
     () =>
