--- src/renderer/views/Login/Login.module.css
@@ -1,18 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.loginBox {
-  padding: 20px;
-  border-radius: 10px;
-  border: 1px solid #ccc;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
-}
-
-.signupContainer {
-  margin-top: 10px;
-  text-align: right;
-}
