--- src/renderer/views/Register/Register.module.css
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.signupBox {
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
-}
-
-.signupContainer {
-  margin-top: 10px;
-  text-align: right;
-}
